+++ packages/__tests__/project1/rolldown.config.ts
@@ -1,3 +1,3 @@
 import { defineConfig } from '@bemedev/dev-utils/rolldown';
 
-export default defineConfig.bemedev();
+export default defineConfig.fast();