+++ packages/core/src/rolldown-config/output.ts
@@ -1,14 +1,17 @@
 import type { OutputOptions } from 'rolldown';
 
 import { DEFAULT_DIR } from './constants';
+import { getConfig } from './plugins/typescript.config';
 
 export const buildOutput = (dir = DEFAULT_DIR, sourcemap: boolean) => {
+  const config = getConfig(process.cwd());
+  const outDir = config.options.outDir ?? dir ?? DEFAULT_DIR;
   return [
     {
       format: 'es',
       sourcemap,
       preserveModulesRoot: 'src',
-      dir,
+      dir: outDir,
       preserveModules: true,
       entryFileNames: '[name].js',
     },
@@ -16,7 +19,7 @@ export const buildOutput = (dir = DEFAULT_DIR, sourcemap: boolean) => {
       format: 'cjs',
       sourcemap,
       preserveModulesRoot: 'src',
-      dir,
+      dir: outDir,
       preserveModules: true,
       entryFileNames: '[name].cjs',
     },