+++ packages/core/src/utils/bemedev/globals/types.ts
@@ -1,4 +1,5 @@
 export type Fn<Args extends any[] = any[], R = any> = (...args: Args) => R;
+export type FnBasic<Main extends Fn, Tr extends object> = Tr & Main;
 
 export type RuA = readonly unknown[];