+++ packages/core/src/utils/bemedev/globals/utils/expandFn.ts
@@ -1,0 +1,17 @@
+import type { Fn, FnBasic } from '../types';
+
+export const expandFn = <
+  Main extends Fn,
+  const Tr extends object = object,
+>(
+  main: Main,
+  extensions?: Tr,
+): FnBasic<Main, Tr> => {
+  const out: any = main;
+
+  if (extensions) {
+    Object.assign(out, extensions);
+  }
+
+  return out;
+};