+++ packages/core/src/utils/index.ts
@@ -2,3 +2,4 @@ export * from './identity';
 export * from './sleep';
 export * from './types';
 export * from './array';
+export * from '#bemedev/globals/utils/_unknown';