+++ packages/core/src/utils/types.ts
@@ -1,5 +1,3 @@
-export type Fn<Args extends unknown[], Return = void> = (
-  ...args: Args
-) => Return;
+export * from '#bemedev/globals/types';
 
 export type ToArray_F = <T>(value?: T | T[]) => T[];